rename :: token -> refresh
