+++ src/service/auth.ts
@@ -1,6 +1,6 @@
 import { signUp } from './auth/signup';
 import { login } from './auth/login';
-import { refresh } from './auth/token';
+import { refresh } from './auth/refresh';
 import { logout } from './auth/logout';
 
 export default {
